Revert "Implement filterBy on DS.ManyArray"
This reverts commit 2f1b54239321a1736011eec5e75cc5596c1d00ce.
